# ci: Better job parallelism, cancel pipeline on failure and extra clippy checks