--- a/.github/workflows/rust.yml
+++ b/.github/workflows/rust.yml
@@ -46,10 +46,38 @@ jobs:
           command: fmt
           args: --all -- --check
 
+  machete:
+    name: "Check unused dependencies"
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout sources
+        uses: actions/checkout@v4
+
+      - name: Use substrate-node binary
+        uses: ./.github/workflows/actions/use-substrate
+
+      - name: Install Rust stable toolchain
+        uses: actions-rs/toolchain@v1
+        with:
+          profile: minimal
+          toolchain: stable
+          override: true
+
+      - name: Rust Cache
+        uses: Swatinem/rust-cache@23bce251a8cd2ffc3c1075eaa2367cf899916d84 # v2.7.3
+
+      - name: Install cargo-machete
+        run: cargo install cargo-machete
+
+      - name: Check unused dependencies
+        uses: actions-rs/cargo@v1.0.3
+        with:
+          command: machete
+
   clippy:
     name: Cargo clippy
     runs-on: ubuntu-latest
-    needs: fmt
+    needs: [fmt, machete]
     steps:
       - name: Checkout sources
         uses: actions/checkout@v4
@@ -77,7 +105,7 @@ jobs:
   check:
     name: Cargo check
     runs-on: ubuntu-latest
-    needs: [fmt, clippy]
+    needs: [fmt, machete]
     steps:
       - name: Checkout sources
         uses: actions/checkout@v4
@@ -136,7 +164,7 @@ jobs:
   wasm_check:
     name: Cargo check (WASM)
     runs-on: ubuntu-latest
-    needs: [fmt, clippy]
+    needs: [fmt, machete]
     steps:
       - name: Checkout sources
         uses: actions/checkout@v4
@@ -157,39 +185,10 @@ jobs:
         run: |
           cargo check --manifest-path examples/wasm-example/Cargo.toml --target wasm32-unknown-unknown
 
-  machete:
-    name: "Check unused dependencies"
-    runs-on: ubuntu-latest
-    needs: [check, wasm_check]
-    steps:
-      - name: Checkout sources
-        uses: actions/checkout@v4
-
-      - name: Use substrate-node binary
-        uses: ./.github/workflows/actions/use-substrate
-
-      - name: Install Rust stable toolchain
-        uses: actions-rs/toolchain@v1
-        with:
-          profile: minimal
-          toolchain: stable
-          override: true
-
-      - name: Rust Cache
-        uses: Swatinem/rust-cache@23bce251a8cd2ffc3c1075eaa2367cf899916d84 # v2.7.3
-
-      - name: Install cargo-machete
-        run: cargo install cargo-machete
-
-      - name: Check unused dependencies
-        uses: actions-rs/cargo@v1.0.3
-        with:
-          command: machete
-
   docs:
     name: Check documentation and run doc tests
     runs-on: ubuntu-latest
-    needs: [check, wasm_check]
+    needs: [fmt, machete]
     steps:
       - name: Checkout sources
         uses: actions/checkout@v4
@@ -219,7 +218,7 @@ jobs:
   tests:
     name: "Test (Native)"
     runs-on: ubuntu-latest-16-cores
-    needs: [machete, docs]
+    needs: [clippy, check, wasm_check, docs]
     timeout-minutes: 30
     steps:
       - name: Checkout sources
@@ -250,7 +249,7 @@ jobs:
   unstable_backend_tests:
     name: "Test (Unstable Backend)"
     runs-on: ubuntu-latest-16-cores
-    needs: [machete, docs]
+    needs: [clippy, check, wasm_check, docs]
     timeout-minutes: 30
     steps:
       - name: Checkout sources
@@ -281,7 +280,7 @@ jobs:
   light_client_tests:
     name: "Test (Light Client)"
     runs-on: ubuntu-latest
-    needs: [machete, docs]
+    needs: [clippy, check, wasm_check, docs]
     timeout-minutes: 15
     steps:
       - name: Checkout sources
@@ -309,7 +308,7 @@ jobs:
   wasm_tests:
     name: Test (WASM)
     runs-on: ubuntu-latest
-    needs: [machete, docs]
+    needs: [clippy, check, wasm_check, docs]
     timeout-minutes: 30
     env:
       # Set timeout for wasm tests to be much bigger than the default 20 secs.