--- a/.github/workflows/rust.yml
+++ b/.github/workflows/rust.yml
@@ -45,6 +45,21 @@ jobs:
         with:
           command: fmt
           args: --all -- --check
+      - if: failure()
+        name: Check Job Status
+        uses: actions/github-script@v6
+        env:
+          RUN_ID: ${{ github.run_id }}
+          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+        with:
+          script: |
+            const runId = process.env.RUN_ID
+            const [owner, repo] = process.env.GITHUB_REPOSITORY.split("/");
+            const resp = await github.rest.actions.cancelWorkflowRun({
+              owner,
+              repo,
+              runId
+            })
 
   machete:
     name: "Check unused dependencies"
@@ -73,6 +88,21 @@ jobs:
         uses: actions-rs/cargo@v1.0.3
         with:
           command: machete
+      - if: failure()
+        name: Check Job Status
+        uses: actions/github-script@v6
+        env:
+          RUN_ID: ${{ github.run_id }}
+          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+        with:
+          script: |
+            const runId = process.env.RUN_ID
+            const [owner, repo] = process.env.GITHUB_REPOSITORY.split("/");
+            const resp = await github.rest.actions.cancelWorkflowRun({
+              owner,
+              repo,
+              runId
+            })
 
   clippy:
     name: Cargo clippy
@@ -101,6 +131,21 @@ jobs:
         with:
           command: clippy
           args: --all-targets -- -D warnings
+      - if: failure()
+        name: Check Job Status
+        uses: actions/github-script@v6
+        env:
+          RUN_ID: ${{ github.run_id }}
+          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+        with:
+          script: |
+            const runId = process.env.RUN_ID
+            const [owner, repo] = process.env.GITHUB_REPOSITORY.split("/");
+            const resp = await github.rest.actions.cancelWorkflowRun({
+              owner,
+              repo,
+              runId
+            })
 
   check:
     name: Cargo check
@@ -160,6 +205,21 @@ jobs:
       # Check the parachain-example code, which isn't a part of the workspace so is otherwise ignored.
       - name: Cargo check parachain-example
         run: cargo check --manifest-path examples/parachain-example/Cargo.toml
+      - if: failure()
+        name: Check Job Status
+        uses: actions/github-script@v6
+        env:
+          RUN_ID: ${{ github.run_id }}
+          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+        with:
+          script: |
+            const runId = process.env.RUN_ID
+            const [owner, repo] = process.env.GITHUB_REPOSITORY.split("/");
+            const resp = await github.rest.actions.cancelWorkflowRun({
+              owner,
+              repo,
+              runId
+            })
 
   wasm_check:
     name: Cargo check (WASM)
@@ -184,6 +244,21 @@ jobs:
       - name: Cargo check WASM examples
         run: |
           cargo check --manifest-path examples/wasm-example/Cargo.toml --target wasm32-unknown-unknown
+      - if: failure()
+        name: Check Job Status
+        uses: actions/github-script@v6
+        env:
+          RUN_ID: ${{ github.run_id }}
+          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+        with:
+          script: |
+            const runId = process.env.RUN_ID
+            const [owner, repo] = process.env.GITHUB_REPOSITORY.split("/");
+            const resp = await github.rest.actions.cancelWorkflowRun({
+              owner,
+              repo,
+              runId
+            })
 
   docs:
     name: Check documentation and run doc tests
@@ -214,6 +289,21 @@ jobs:
         with:
           command: test
           args: --doc
+      - if: failure()
+        name: Check Job Status
+        uses: actions/github-script@v6
+        env:
+          RUN_ID: ${{ github.run_id }}
+          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+        with:
+          script: |
+            const runId = process.env.RUN_ID
+            const [owner, repo] = process.env.GITHUB_REPOSITORY.split("/");
+            const resp = await github.rest.actions.cancelWorkflowRun({
+              owner,
+              repo,
+              runId
+            })
 
   tests:
     name: "Test (Native)"
@@ -245,6 +335,21 @@ jobs:
         with:
           command: nextest
           args: run --workspace
+      - if: failure()
+        name: Check Job Status
+        uses: actions/github-script@v6
+        env:
+          RUN_ID: ${{ github.run_id }}
+          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+        with:
+          script: |
+            const runId = process.env.RUN_ID
+            const [owner, repo] = process.env.GITHUB_REPOSITORY.split("/");
+            const resp = await github.rest.actions.cancelWorkflowRun({
+              owner,
+              repo,
+              runId
+            })
 
   unstable_backend_tests:
     name: "Test (Unstable Backend)"
@@ -276,6 +381,21 @@ jobs:
         with:
           command: nextest
           args: run --workspace --features unstable-backend-client
+      - if: failure()
+        name: Check Job Status
+        uses: actions/github-script@v6
+        env:
+          RUN_ID: ${{ github.run_id }}
+          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+        with:
+          script: |
+            const runId = process.env.RUN_ID
+            const [owner, repo] = process.env.GITHUB_REPOSITORY.split("/");
+            const resp = await github.rest.actions.cancelWorkflowRun({
+              owner,
+              repo,
+              runId
+            })
 
   light_client_tests:
     name: "Test (Light Client)"
@@ -304,6 +424,21 @@ jobs:
         with:
           command: test
           args: --release --package integration-tests --features unstable-light-client
+      - if: failure()
+        name: Check Job Status
+        uses: actions/github-script@v6
+        env:
+          RUN_ID: ${{ github.run_id }}
+          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+        with:
+          script: |
+            const runId = process.env.RUN_ID
+            const [owner, repo] = process.env.GITHUB_REPOSITORY.split("/");
+            const resp = await github.rest.actions.cancelWorkflowRun({
+              owner,
+              repo,
+              runId
+            })
 
   wasm_tests:
     name: Test (WASM)
@@ -357,3 +492,18 @@ jobs:
           wasm-pack test --headless --firefox
           wasm-pack test --headless --chrome
         working-directory: signer/wasm-tests
+      - if: failure()
+        name: Check Job Status
+        uses: actions/github-script@v6
+        env:
+          RUN_ID: ${{ github.run_id }}
+          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+        with:
+          script: |
+            const runId = process.env.RUN_ID
+            const [owner, repo] = process.env.GITHUB_REPOSITORY.split("/");
+            const resp = await github.rest.actions.cancelWorkflowRun({
+              owner,
+              repo,
+              runId
+            })