--- a/.github/workflows/rust.yml
+++ b/.github/workflows/rust.yml
@@ -130,7 +130,38 @@ jobs:
         uses: actions-rs/cargo@v1
         with:
           command: clippy
-          args: -p subxt --no-default-features --features web,unstable-light-client,jsonrpsee --target wasm32-unknown-unknown   -- -D warnings
+          args: -p subxt --no-default-features --features web,unstable-light-client,jsonrpsee --target wasm32-unknown-unknown -- -D warnings
+
+      - if: "failure()"
+        uses: "andymckay/cancel-action@0.3"
+
+  wasm_clippy:
+    name: Cargo clippy (WASM)
+    runs-on: ubuntu-latest
+    needs: [fmt, machete]
+    steps:
+      - name: Checkout sources
+        uses: actions/checkout@v4
+
+      - name: Use substrate-node binary
+        uses: ./.github/workflows/actions/use-substrate
+
+      - name: Install Rust stable toolchain
+        uses: actions-rs/toolchain@v1
+        with:
+          profile: minimal
+          toolchain: stable
+          target: wasm32-unknown-unknown
+          override: true
+
+      - name: Rust Cache
+        uses: Swatinem/rust-cache@23bce251a8cd2ffc3c1075eaa2367cf899916d84 # v2.7.3
+
+      - name: Run clippy
+        uses: actions-rs/cargo@v1
+        with:
+          command: clippy
+          args: -p subxt --no-default-features --features web,unstable-light-client,jsonrpsee --target wasm32-unknown-unknown -- -D warnings
 
       - if: "failure()"
         uses: "andymckay/cancel-action@0.3"
@@ -260,7 +291,7 @@ jobs:
   tests:
     name: "Test (Native)"
     runs-on: ubuntu-latest-16-cores
-    needs: [clippy, check, wasm_check, docs]
+    needs: [clippy, wasm_clippy, check, wasm_check, docs]
     timeout-minutes: 30
     steps:
       - name: Checkout sources
@@ -294,7 +325,7 @@ jobs:
   unstable_backend_tests:
     name: "Test (Unstable Backend)"
     runs-on: ubuntu-latest-16-cores
-    needs: [clippy, check, wasm_check, docs]
+    needs: [clippy, wasm_clippy, check, wasm_check, docs]
     timeout-minutes: 30
     steps:
       - name: Checkout sources
@@ -328,7 +359,7 @@ jobs:
   light_client_tests:
     name: "Test (Light Client)"
     runs-on: ubuntu-latest
-    needs: [clippy, check, wasm_check, docs]
+    needs: [clippy, wasm_clippy, check, wasm_check, docs]
     timeout-minutes: 15
     steps:
       - name: Checkout sources
@@ -359,7 +390,7 @@ jobs:
   wasm_tests:
     name: Test (WASM)
     runs-on: ubuntu-latest
-    needs: [clippy, check, wasm_check, docs]
+    needs: [clippy, wasm_clippy, check, wasm_check, docs]
     timeout-minutes: 30
     env:
       # Set timeout for wasm tests to be much bigger than the default 20 secs.