There are 3 main changes included in this PR:
- The CI pipeline is shuffled for better parallelism and faster build times
- Cancelling the pipeline if one of the CI steps fails (as tested by [this job](https://github.com/paritytech/subxt/actions/runs/7709721254/job/21011517365))
- Introduce better clippy for steps to account for `unstable-light-client` and `web` feature flags
  - As a result of this, the clippy warnings are also fixed

Change the CI parallelism for faster build times:

### Step 1
- fmt, machete

### Step 2
- clippy, check, wasm_check, docs

### Step 3
- all other tests


## Time Saved

We are saving an average of 10min per successful CI time. This is the time it takes the CI from starting the first job to finishing the last job, not the total amount of CI minutes used.

On average, the CI minutes have increased by 39sec because we introduced another clippy check for WASM.

```
Before:
    6s + 3m54s + max(5m50s, 1m34s) + max(16s, 5m50s) + max(9m27s, 9m24s, 6m2s, 6m10s) = 1507s (25.12 min)

After:
    max(18s, 15s) + max(5m24s, 39s, 6m3s, 1m25s, 5m53s) + max(9m43s, 9m21s, 6m7s, 6m36s) = 954s (15.9 min)
```

### Before

Taken from https://github.com/paritytech/subxt/actions/runs/7710633300.
![before](https://github.com/paritytech/subxt/assets/60601340/34cd35d8-1197-4c7f-8ce9-e21f1d92b516)

### After

Taken from https://github.com/paritytech/subxt/actions/runs/7714660920. 

![after](https://github.com/paritytech/subxt/assets/60601340/39690fea-092a-40c2-8107-daf86402d3e7)